+++ src/dashboard/CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to
 - add mass admin action (make revoked and make awaiting) for consents
 - block the updates of all new data if a consent has the status `VALIDATED`
 - block the deletion of consent if it has the status `VALIDATED`
+- block consent updates (via the consent form) if the consent status is not `AWAITING`
 - integration of custom 403, 404 and 500 pages 
 - sentry integration
 - added a signal on the creation of a delivery point. This signal allows the creation